-- Arcade -- Fixed bug causing TNT to not be handled by the plugin, this was the reason behind the maps being destroyed by explosions.
Fixed bug causing getLastCombat() to error out and give no response, it will not return null if the player has not been in combat.

Init function has been optimized, and moved less important operations to bottom of the code.

Cleanup will now finish in 3 seconds instead of 1, this will allow more in-depth operations to be executed.